Load empty mo file only once
This reduces the amount of reads when translating strings for the
default locale. Default locales it does not have a po file to load and
fall back to an empty mo file loading `empty.mo` each time.

Previously, the following calls result in reading the empty mo file over
and over again:

  # default locale - no translations available/loaded
  _('foo') # => reads empty mo file
  _('foo') # => second call is then cached
  _('bar') # => reads empty mo file
  _('baz') # => reads empty mo file

Now, it's:

  # default locale - no translations available/loaded
  _('foo') # => reads empty mo file
  _('foo') # => second call is then cached
  _('bar') # => uses ivar cache
  _('baz') # => uses ivar cache