Add explicit dependency to `prime`
The `prime` gem has been extracted from Ruby core into a gem.

Starting with Ruby 3.4.0, it will no longer be bundled with Ruby, so it requires an explicit dependency. Ruby 3.3.0+ display a warning due to this:

```
/usr/local/bundle/gems/bootsnap-1.18.3/lib/bootsnap/load_path_cache/core_ext/kernel_require.rb:27: warning: prime was loaded from the standard library, but is not part of the default gems since Ruby 3.1.0. Add prime to your Gemfile or
gemspec. Also contact author of fast_gettext-2.3.0 to add prime into its gemspec.
```

I also removed the fallback that uses `mathn`, as `mathn` is deprecated and not recommended to use. The `prime` gem supports Ruby 2.5.0+, same as this gem, so it should not break compatibility.