Convert String#% to Ruby 1.9+ behavior with option use maintain current behavior
Previously String#% was overridden on versions after Ruby 1.9 to match the 1.8 behavior where missing keys in the expanded format are ignored and not replaced. This is likely suprising behavior in modern Ruby. This removes the previous patch, but still allows overriding this behavior optionally by calling `FastGettext.allow_invalid_keys!`.